ändrad struktur för seeding ingredients
