--- backend/src/features/ingredients/data/ingredientSeed.ts
@@ -1,8 +1,0 @@
-export const ingredientSeed = [
-  { name: "Tunnbröd", stock: 100 },
-  { name: "Renkött", stock: 100 },
-  { name: "Kål", stock: 100 },
-  { name: "Rostat bovete", stock: 100 },
-  { name: "Blåbärsdressing", stock: 100 }
-];
-
