+++ backend/src/features/ingredients/data/ingredientsSeed.ts
@@ -1,0 +1,42 @@
+export const ingredientsSeed = [
+    { name: "renkött", stock: 100 },
+    { name: "tunnbröd", stock: 100 },
+    { name: "kål", stock: 100 },
+    { name: "viltskav", stock: 100 },
+    { name: "viltkorv", stock: 100 },
+    { name: "älg", stock: 100 },
+    { name: "svamp", stock: 100 },
+    { name: "lingon", stock: 100 },
+    { name: "potatis", stock: 100 },
+    { name: "lök", stock: 100 },
+    { name: "rödlök", stock: 100 },
+    { name: "blåbär", stock: 100 },
+    { name: "dryck chai", stock: 100 },
+    { name: "dryck varm choklad", stock: 100 },
+    { name: "citron", stock: 100 },
+    { name: "dipp blåbär och chili", stock: 100 },
+    { name: "dipp tallskott och citron", stock: 100 },
+    { name: "dipp rökt svamp", stock: 100 },
+    { name: "rotchips", stock: 100 },
+    { name: "dipp örtdip", stock: 100 },
+    { name: "havregrädde", stock: 100 },
+    { name: "morot", stock: 100 },
+    { name: "kantareller", stock: 100 },
+    { name: "äpple", stock: 100 },
+    { name: "rödbetsburgare", stock: 100 },
+    { name: "portabelloburgare", stock: 100 },
+    { name: "ost", stock: 100 },
+    { name: "vitlök", stock: 100 },
+    { name: "bröd", stock: 100 },
+    { name: "ris", stock: 100 },
+    { name: "bönor", stock: 100 },
+    { name: "gurka", stock: 100 },
+    { name: "hjort", stock: 100 },
+    { name: "grillade rotsaker", stock: 100 },
+    { name: "dryck skogsbärsslush", stock: 100 },
+    { name: "picklad lök", stock: 100 },
+    { name: "tranbärs-chutney", stock: 100 },
+    { name: "granris-majo", stock: 100 },
+    { name: "dryck tallskott och citron", stock: 100 },
+  ];
+
